LPS-76193 Use activator to initialize KB service release number
