+++ modules/apps/knowledge-base/knowledge-base-service/bnd.bnd
@@ -1,3 +1,4 @@
+Bundle-Activator: com.liferay.knowledge.base.internal.activator.KnowledgeBaseServiceBundleActivator
 Bundle-Name: Liferay Knowledge Base Service
 Bundle-SymbolicName: com.liferay.knowledge.base.service
 Bundle-Version: 1.3.20
