+++ modules/apps/knowledge-base/knowledge-base-service/build.gradle
@@ -23,6 +23,7 @@ dependencies {
 	provided group: "com.liferay.portal", name: "com.liferay.util.java", version: "2.0.0"
 	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
 	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
+	provided group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided group: "rome", name: "rome", version: "1.0"
 	provided project(":apps:collaboration:subscription:subscription-api")
